requirements_doc.txt
gemmi==0.5.3
ispyb==6.10.0
matplotlib==3.5.1
mrcfile==1.3
pandas==1.4.2
pillow==9.2.0
plotly==5.7.0
Sphinx==4.5.0
sphinx-rtd-theme==1.0.0
